Plotting tools adjustments
- PDF plots are now produced only for the first q2 value by default.
  New option --q2all to get PDF plots at all q2 values

- Chi2 and parameters tables are produced and added to the plots.pdf file by default.
  New option --no-tables to desactivate tables

- Fit status (converged, not pos-def) is added to the parameters table. If fit error matrix
  is not positive definite, parameters error are displayed in red.

- Added p-value to chi2 table

- New option --chi2-nopdf-uncertainties
  When chi2 is evaluated with the LHAPDFError routine, this option will add to the chi2 table 
  the chi2 evaluated without PDF uncertainties within brackets

- Cleanup of latex auxiliary and log files

- Data plots renamed to data_<dataindex><subplotindex> (before it was <dataindex><subplotindex>_pulls)

- Check that output directory are not given to command line options more than once (avoid segmentation fault)

- Redraw data markers over theory uncertainty bands

- Keep single pages of plots.pdf if any of the splitplots option is requested

- Improved number of divisions in y axis of PDF ratio plots



git-svn-id: https://svnsrv.desy.de/desy/h1fitter/trunk@1445 b2a8d302-ca0a-427a-b7b9-081adbce06bc

+++ tools/DrawResultsPackage/src/DrawPdfs.cc
@@ -109,16 +109,20 @@ int main(int argc, char **argv)
 	}
       //  vector <TCanvas*> pdfscanvaslist, pdfscanvasratiolist;
       for (map <float, pdfmap>::iterator qit = q2list.begin(); qit != q2list.end(); qit++)
-	for (pdfmap::iterator pdfit = (*qit).second.begin(); pdfit != (*qit).second.end(); pdfit++)
-	  {
-	    pdfscanvaslist.push_back(PdfsPainter((*qit).first, (*pdfit).first, (*pdfit).second));
-	    if (info_output.size() > 1)
-	      pdfscanvasratiolist.push_back(PdfsRatioPainter((*qit).first, (*pdfit).first, (*pdfit).second));
-	  }
+	{
+	  for (pdfmap::iterator pdfit = (*qit).second.begin(); pdfit != (*qit).second.end(); pdfit++)
+	    {
+	      pdfscanvaslist.push_back(PdfsPainter((*qit).first, (*pdfit).first, (*pdfit).second));
+	      if (info_output.size() > 1)
+		pdfscanvasratiolist.push_back(PdfsRatioPainter((*qit).first, (*pdfit).first, (*pdfit).second));
+	    }
+	  if (!opts.q2all)
+	    break;
+	}
     }
 
   //--------------------------------------------------
-  //Data pulls plots
+  //Data plots
   gStyle->SetEndErrorSize(4);
   map <int, vector<dataseth> > datamap;
   if (! opts.nodata)
@@ -160,7 +164,7 @@ int main(int argc, char **argv)
 		      if (!bincenter)
 			{
 			  cout << "bin inconsistency for dataset: " << info_output[o]->GetSet(d)->GetName() << " Subplot " << p << endl;
-			  cout << "Cannot plot data pulls, skipping" << endl;
+			  cout << "Cannot plot data, skipping" << endl;
 			  continue;
 			}
 		    }
@@ -190,6 +194,10 @@ int main(int argc, char **argv)
     shiftcanvaslist = ShiftPainter(opts.dirs);
 
   //--------------------------------------------------
+  //Create output directory
+  system(((string)"mkdir -p " + opts.outdir).c_str());
+
+  //--------------------------------------------------
   //Fit and parameters results plots
   if (! opts.notables)
     {
@@ -198,7 +206,6 @@ int main(int argc, char **argv)
     }
 
   //Save plots
-  system(((string)"mkdir -p " + opts.outdir).c_str());
 
   int pgn = 0;
   char pgnum[15];
@@ -388,8 +395,9 @@ int main(int argc, char **argv)
   cout << "Plots saved in: " << (opts.outdir + "plots." + format) << endl;
 
   //cleanup pages
-  if (pgn > 0)
-    system(("rm " + inputfiles).c_str());
+  if (!opts.splitplots)
+    if (pgn > 0)
+      system(("rm " + inputfiles).c_str());
 
   return 0;
 }
